New requests to server alright
Returning a list of variants (SNPs, Indel or both) located in a specific region of a specified chromosome in a specific dataset.

listing the number of variants (SNPs, Indel, both) contained in ecah genome, grouped by chromosome

Left to do: plots of variants density for a specified window size across a specified chromosome, for a specific dataset.